Add host1x bus support
This bus is exclusively used by Tegra devices.  Why this isn’t part of
the platform bus, I do not know.  It has the exact same properties as
the platform bus, so it is pretty much a copy/paste of it, with just
different structs in xf86drm.h.

This has been tested on a Nintendo Switch running Linux 6.0.